requirements.txt
langchain
openai
faiss-cpu
streamlit
pypdf
tiktoken
python-dotenv
langchain-community
langchain-openai